+++ tools/make.cfg
@@ -1,0 +1,102 @@
+# EXAMPLE MAKE.CFG FILE
+# Please see the comments for each option. Most options can be left
+# off for sane defaults.
+
+#################################################################
+# Default make target
+#################################################################
+[DEFAULT]
+
+# Project name (with @ symbol)
+# This is used for naming the release files.
+# Default: Current folder name
+project = @cba
+
+# Path to project secret key for signing
+# Make sure this isn't in your public repository!
+# Default: <work_drive>\private_keys\cba_2.0.0.biprivatekey
+# key = P:\private_keys\cba_2.0.0.biprivatekey
+
+# Path to where private keys are automatically created if the command-line parameter "key" is used 
+# Make sure this isn't in your public repository!
+# Default: <work_drive>\private_keys
+# private_key_path = P:\private_keys
+
+# If set to True, the make system will attempt to autodetect addons in the
+# current folder by looking for directories with 'config.cpp' in them.
+# Default: True
+# module_autodetect = True
+
+# List of directories to ignore when autodetecting addons.
+# Default: release
+# ignore = release, my_unfinished_module
+
+# If autodetect is set to False, only folders whose names are in this list
+# will be built as modules.
+# Default: None
+# modules = my_module, my_supporting_module
+
+# This is the folder hierarchy that will be used as prefix inside the PBO.
+# Default: None
+prefix = x\cba\addons
+
+# Set the location where the addon source folders (i.e. P:\x\cba\addons)
+# Default: <work_drive>\<prefix>\addons
+module_root = P:\x\cba\addons
+
+# Set the location where the optional addon source folders (i.e. P:\z\ace\optionals)
+# Default: <work_drive>\<prefix>\optionals
+optionals_root = P:\z\ace\optionals
+
+
+# If the command-line variable test, the addons built will be copied to the following folder.
+# Default: %USERPROFILE%\documents\Arma 3\<project>
+# test_dir = %USERPROFILE%\documents\Arma 3\<project>
+
+
+# Directory where the built addon will be saved.
+# Default: release
+release_dir = P:\x\cba\release
+
+# This string will be prefixed to all build PBO file names.
+# Default: None
+pbo_name_prefix = cba_
+
+# This string will be prefixed to release archive.
+# Default: None
+zipPrefix = cba
+
+# Which build tool will be used? Options: pboproject, addonbuilder
+# Default: addonbuilder
+build_tool = pboproject
+
+
+##################################################################
+# Optional CBA build 
+###################################################################
+[Optional]
+prefix = x\cba\addons
+module_root = P:\x\cba\optional
+
+
+
+##################################################################
+# Alternate build target using a different key
+###################################################################
+# [DifferentKey]
+# key = C:\Keys\different.biprivatekey
+
+##################################################################
+# Alternate build target ignoring some modules when detecting
+###################################################################
+# [IgnoreSome]
+# key = C:\Keys\different.biprivatekey
+# ignore = release, my_server_module, my_private_module
+
+##################################################################
+# Alternate build target with fixed build list
+###################################################################
+# [Fixed]
+# module_autodetect = False
+# modules = my_module, my_other_module
+
