+++ tools/make_functions_config.bat
@@ -1,0 +1,3 @@
+ruby functions_config.rb "../addons"
+
+pause
