--- utils/make_functions_config.bat
@@ -1,3 +1,0 @@
-ruby functions_config.rb "../addons"
-
-pause
