--- utils/ndocs-project/cba.css
@@ -1,11 +1,0 @@
-pre {
-	margin: 0;
-	background: #EEEEEE;
-	border: 1px solid #888888;
-	padding: 0.5em 1em 0.5em 1em;
-}
-
-p {
-	text-indent: 0;
-	margin-bottom: 1em;
-}
