--- utils/ndocs-project/menu_template.txt
@@ -1,80 +1,0 @@
-Format: 1.4
-
-
-Title: CBA
-SubTitle: Community Base Addons
-
-Footer: Copyright (c) CBA team and contributers
-
-Timestamp: year
-
-#   m     - One or two digit month.  January is "1"
-#   mm    - Always two digit month.  January is "01"
-#   mon   - Short month word.  January is "Jan"
-#   month - Long month word.  January is "January"
-#   d     - One or two digit day.  1 is "1"
-#   dd    - Always two digit day.  1 is "01"
-#   day   - Day with letter extension.  1 is "1st"
-#   yy    - Two digit year.  2006 is "06"
-#   yyyy  - Four digit year.  2006 is "2006"
-#   year  - Four digit year.  2006 is "2006"
-
-# These are indexes you deleted, so Natural Docs will not add them again
-# unless you remove them from this line.
-
-Don't Index: Files
-
-
-# --------------------------------------------------------------------------
-# 
-# Cut and paste the lines below to change the order in which your files
-# appear on the menu.  Don't worry about adding or removing files, Natural
-# Docs will take care of that.
-# 
-# You can further organize the menu by grouping the entries.  Add a
-# "Group: [name] {" line to start a group, and add a "}" to end it.
-# 
-# You can add text and web links to the menu by adding "Text: [text]" and
-# "Link: [name] ([URL])" lines, respectively.
-# 
-# The formatting and comments are auto-generated, so don't worry about
-# neatness when editing the file.  Natural Docs will clean it up the next
-# time it is run.  When working with groups, just deal with the braces and
-# forget about the indentation and comments.
-# 
-# --------------------------------------------------------------------------
-
-File: Overview (no auto-title, overview.txt)
-
-Index: Index (Everything)
-	
-Group: Project  {
-	Link: Wiki  (https://github.com/CBATeam/CBA_A3/wiki)
-	Link: Reference (http://cbateam.github.io/CBA_A3/docs/files/overview-txt.html)
-	Link: Forums  (http://forums.bistudio.com/showthread.php?178224-CBA-Community-Base-Addons-ARMA-3)
-	Link: Issues  (https://github.com/CBATeam/CBA_A3/issues)
-	Link: Download  (https://github.com/CBATeam/CBA_A3/releases)
-	}  # Group: Project
-   
-	File: Overview (doc_overview.txt)
-	  
-
-	
-Group: Functions  {
-
-	Function Index: Index
-   
-$FUNCTIONS$
-
-	}  # Group: Functions
-
-Group: Macros  {
-
-	Macro Index: Index
-		
-	File: script_macros_common (no auto-title, main\script_macros_common.hpp)
-
-	}  # Group: Macros
-   
- 
-
