--- utils/skel/script_component.hpp
@@ -1,12 +1,0 @@
-#define COMPONENT __COMPONENT__
-#include "\x\__PREFIX__\addons\main\script_mod.hpp"
-
-#ifdef DEBUG_ENABLED___COMPONENTU__
-	#define DEBUG_MODE_FULL
-#endif
-
-#ifdef DEBUG_SETTINGS___COMPONENTU__
-	#define DEBUG_SETTINGS DEBUG_SETTINGS___COMPONENTU__
-#endif
-
-#include "\x\__PREFIX__\addons\main\script_macros.hpp"
